refactor(contracts): extract OperationContext
Step 2 of moving acp/managed_agent out of kernel. #3932's author
confirmed they don't touch OperationContext, so this lands cleanly
ahead of their merge -- no rebase conflict either way.

Once #3932 lands and the services crate gains kernel-tier dep
permission (already true today per #3932 author's confirmation:
"services ⊥ backends ⊥ transport ⊥ raft are peer-crate
non-deps but each peer can depend on kernel + contracts + lib"),
acp/managed_agent will move to rust/services/src/{acp,managed_agent}/
and need to construct system-tier OperationContext values for
sys_read/sys_write/sys_unlink calls. Having the type in contracts
means services::acp can build it without taking kernel as a
direct construction-site dep.

* rust/contracts/src/operation_context.rs  -- new (~70 LOC)
* rust/contracts/src/lib.rs                 -- pub mod + re-export
* rust/kernel/src/kernel.rs                 -- delete the local
                                               struct, re-export
                                               from contracts so
                                               every crate::kernel::
                                               OperationContext call
                                               site stays unchanged

Verification:
  cargo clippy -p kernel --tests --all-features -- -D warnings  clean
  cargo test -p kernel --lib                  596 passed, 0 failed
